+++ frontend/src/metabase/plugins/oss/embedding-iframe-sdk.ts
@@ -1,0 +1,6 @@
+import type { ReactNode } from "react";
+
+export const PLUGIN_EMBEDDING_IFRAME_SDK = {
+  hasValidLicense: () => false,
+  SdkIframeEmbedRoute: (): ReactNode => null,
+};
