+++ frontend/src/metabase/plugins/oss/embedding-sdk.ts
@@ -1,0 +1,6 @@
+export const PLUGIN_EMBEDDING_SDK = {
+  isEnabled: () => false,
+  onBeforeRequestHandlers: {
+    getOrRefreshSessionHandler: async () => {},
+  },
+};
