+++ frontend/src/metabase/plugins/oss/entities.ts
@@ -1,0 +1,3 @@
+export const PLUGIN_ENTITIES = {
+  entities: {} as Record<string, any>,
+};
