+++ frontend/src/metabase/plugins/oss/public-sharing.ts
@@ -1,0 +1,6 @@
+import type React from "react";
+
+export const PLUGIN_PUBLIC_SHARING = {
+  PublicDocumentRoute: (_props: any) => null as React.ReactElement | null,
+  PublicLinksDocumentListing: () => null as React.ReactElement | null,
+};
