+++ frontend/src/metabase/plugins/oss/semantic-search.ts
@@ -1,0 +1,5 @@
+import { PluginPlaceholder } from "metabase/plugins/components/PluginPlaceholder";
+
+export const PLUGIN_SEMANTIC_SEARCH = {
+  SearchSettingsWidget: PluginPlaceholder,
+};
