+++ frontend/src/metabase/plugins/oss/smtp-override.ts
@@ -1,0 +1,11 @@
+import type { ComponentType } from "react";
+
+import { PluginPlaceholder } from "metabase/plugins/components/PluginPlaceholder";
+
+export const PLUGIN_SMTP_OVERRIDE: {
+  CloudSMTPConnectionCard: ComponentType;
+  SMTPOverrideConnectionForm: ComponentType<{ onClose: () => void }>;
+} = {
+  CloudSMTPConnectionCard: PluginPlaceholder,
+  SMTPOverrideConnectionForm: PluginPlaceholder,
+};
