+++ frontend/src/metabase/plugins/oss/whitelabel.ts
@@ -1,0 +1,6 @@
+import { PluginPlaceholder } from "metabase/plugins/components/PluginPlaceholder";
+
+export const PLUGIN_WHITELABEL = {
+  WhiteLabelBrandingSettingsPage: PluginPlaceholder,
+  WhiteLabelConcealSettingsPage: PluginPlaceholder,
+};
